Change interference from tinyint to int instead of to smallint

+++ database/migrations/2025_02_12_174229_change_access_points_interference_tinyint_to_smallint.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('access_points', function (Blueprint $table) {
-            $table->unsignedSmallInteger('interference')->change();
+            $table->unsignedInteger('interference')->change();
         });
     }
